moved some code into hooks that was essentially being repeated in multiple places
